requirements.txt
openai==1.75.0
python-pptx==1.0.2
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
